[EasyRepository] Create trait for DoctrineOrmRepository to make it easier to use DoctrineBundle ServiceEntityRepository
